app/requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.1
decorator==5.1.1
defusedxml==0.7.1
einops==0.6.1
executing==1.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.18.0
filelock==3.12.2
fonttools==4.40.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
hnswlib==0.7.0
imageio==2.31.1
ipython==8.14.0
jax==0.4.19
jax-jumpy==1.0.0
jaxlib==0.4.19
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapy @ git+https://github.com/PWhiddy/mediapy.git@45101800d4f6adeffe814cad93de1db67c1bd614
mistune==3.0.1
ml-dtypes==0.2.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
networkx==3.1
numpy==1.25.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyboy==1.6.9
Pygments==2.15.1
pyparsing==3.1.0
PySDL2==0.9.15
pysdl2-dll==2.28.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pyzmq==25.1.1
referencing==0.30.2
rpds-py==0.9.2
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Shimmy==1.1.0
six==1.16.0
soupsieve==2.4.1
stable-baselines3==2.0.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.15.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
torch==2.0.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
wandb==0.15.12
wcwidth==0.2.6
webencodings==0.5.1

